[FramePacketizer, MemoryManage, AudioCapture] 프로젝트 분리 및 오디오 캡쳐 예시 추가 - FramePacketizer에서 메모리 관리 코드 분리     프로젝트의 크기를 줄이기 위해 다른 프로젝트로 분리
- 윈도우 오디오 캡쳐 및 인코딩 예시 코드 작성
    윈도우에서 raw 오디오 데이터를 얻는 방식와 해당 데이터를 인코딩 하는 방법을 기록
    추후 캡쳐 클래스와 인코딩 클래스 제작 예정

+++ src/Live555RtspServer/CaptureMediaSubsession.cpp
@@ -20,7 +20,7 @@ void CaptureMediaSubsession::setSPSNAL(const BYTE* ptr, int size)
 	fSPSNALSize = size;
 }
 
-void CaptureMediaSubsession::setBitRate(BYTE br)
+void CaptureMediaSubsession::setBitRate(unsigned long br)
 {
 	if (br > 102400)
 	{
